Update src/c/finduses.nw
In this version, finduses will check for the presence of an environment variable
FINDUSES_LISP set to something other than "0". If this is the case, then
specific lisp syntax will be applied when building the index references and
hyperlinks in html. If not, then the standard syntax will be applied.

http://home.scarlet.be/asld0009/Noweb/autodef.html